requirements.txt
# Web Application
Django==1.8.5
django-braces==1.8.1
django-dotenv==1.3.0
whitenoise==2.0.4
gunicorn==19.3.0

# Database Requirements
psycopg2==2.6.1
dj-database-url==0.3.0

# Utilities
six==1.10.0
wheel==0.24.0
## The following requirements were added by pip freeze:
django-grappelli==2.7.1
python-dateutil==2.4.2
